feat: Add study setup component and notebook type
Introduce a new `StudySetup` component to allow users to configure their study sessions. This includes options for random question selection and selection from specific notebooks.

Also, define a `QuestionNotebook` type to represent collections of questions, enabling structured study. A `CubeIcon` was added to the Icons component for use in the new `StudySetup` component.